Add prototype feed data for demoing propuse
