+++ Prototype/Prototype/Helpers/UIView+Extensions.swift
@@ -1,0 +1,14 @@
+//
+//  UIView+Extensions.swift
+//  Prototype
+//
+//  Created by Mauro Worobiej on 02/05/2022.
+//
+
+import UIKit
+
+extension UIView {
+    func addSubviews(_ views: UIView...) {
+        views.forEach {addSubview($0)}
+    }
+}
